+++ appstack_plugin/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the Appstack Flutter Plugin will be documented in this fi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2025-12-08
+
+### Added
+- New `getAttributionParams()` method to retrieve attribution parameters from the SDK
+- Updated iOS and Android SDK references to latest versions with attribution params support
+- Added attribution params display in sample app
+
 ## [1.0.4] - 2025-12-04
 
 ### Fixed